Split MintConfirmed event, unify some logic, linter, cleanups.
Split MintConfirmed to 2 events: MintConfirmed and MintRejected

Unify some duplicated logic:
-getNonceAndMintRequest
-getTimestamp

Change cancelMintRequest to use struct in memory and not storage,
so it can also use getNonceAndMintRequest.

Change IndexedMapping.indexedMapping to IndexedMapping.Data for proper
capitalization.

Remove ForceEther as we don't need it for testing reclaimToken,
as the latter was removed.

Small indentation changes to satisfy linter.