100% code coverage (including branches).
In order to cover the "unknown/unreachable" return option in Factory's
getStatusString, an "unreachable" enum value was also added (otherwise can not
reach a non existing enum value and resulting with invalid opcode instead of
reaching the desired line).
It seems reasonable, assuming we will not forfeit the 100% code coverage and
do not want to remove the unreachable status string (which is a good practice).
Also getStatusString visibility was changed to public to be able to test it
directly.

Some require error messages were unified to a similar "invalid X address"
phrase.