Fix Dependencies
The tests were not running correctly because solc, truffle, and web3 were using
carat semver, pulling in higher versions that were incompatible with the
repo as it was.

This commit removes the carat for solc, truffle, and web3 in package.json

This commit also adds npm scripts for compiling and running tests. It
update the README accordingly